balloon-animation/requirements.txt
# Required dependencies for Stanford Code in Place final project
graphics.py==5.0  # Custom graphics library from Stanford Code in Place
